Restructured dls package
