+++ bee.persistent.test/src/test/kotlin/com/beeproduced/bee/persistent/test/BeePersistentTestA.kt
@@ -3,13 +3,17 @@ package com.beeproduced.bee.persistent.test
 import com.beeproduced.bee.persistent.application.Application
 import com.beeproduced.bee.persistent.blaze.dsl.entity.Path
 import com.beeproduced.bee.persistent.blaze.dsl.entity.ValuePath
-import com.beeproduced.bee.persistent.blaze.dsl.expression.StringFunctions.lower
-import com.beeproduced.bee.persistent.blaze.dsl.select.and
-import com.beeproduced.bee.persistent.blaze.dsl.select.or
+import com.beeproduced.bee.persistent.blaze.dsl.expression.builder.lower
+import com.beeproduced.bee.persistent.blaze.dsl.predicate.builder.and
+import com.beeproduced.bee.persistent.blaze.dsl.predicate.builder.or
 import com.beeproduced.bee.persistent.blaze.meta.dsl.InlineValueUnwrapper
 import com.beeproduced.bee.persistent.blaze.meta.dsl.InlineValueUnwrappers
+import com.beeproduced.bee.persistent.blaze.selection.BeeSelection
+import com.beeproduced.datasource.a.*
 import jakarta.persistence.EntityManager
+import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import org.junit.jupiter.api.extension.ExtendWith
 import org.springframework.beans.factory.annotation.Autowired
@@ -18,10 +22,6 @@ import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.test.context.junit.jupiter.SpringExtension
 import org.springframework.transaction.PlatformTransactionManager
 import org.springframework.transaction.support.TransactionTemplate
-import com.beeproduced.bee.persistent.blaze.selection.BeeSelection
-import com.beeproduced.datasource.a.*
-import org.junit.jupiter.api.AfterEach
-import org.junit.jupiter.api.Test
 import java.lang.invoke.LambdaMetafactory
 import java.lang.invoke.MethodHandles
 import java.lang.invoke.MethodType
