+++ bee.persistent/src/blaze/kotlin/com/beeproduced/bee/persistent/blaze/dsl/expression/function/string/LowerExpression.kt
@@ -8,7 +8,7 @@ import com.beeproduced.bee.persistent.blaze.dsl.expression.Expression
  * @author Kacper Urbaniec
  * @version 2024-01-11
  */
-class LowerExpression(
+internal class LowerExpression(
     private val expression: Expression<String>
 ) : Expression<String> {
     override fun Expression<String>.toExpressionString(): String
